BusinessPedia/BusinessPedia/requirements.txt
mysql-connector-python
Pillow
playsound 